+++ CMSIS/DSP/Source/StatisticsFunctions/arm_absmin_no_idx_f16.c
@@ -196,7 +196,7 @@ void arm_absmin_no_idx_f16(
 
 
   /* Load first input value that act as reference value for comparision */
-  out = fabsf(*pSrc++);
+  out = (_Float16)fabsf((float32_t)*pSrc++);
 
   /* Initialize blkCnt with number of samples */
   blkCnt = (blockSize - 1U);
